`lt_isolate`
